add data for cynthia boss fight + fix catch mechanic +  replace all pokemons catch rate

+++ app/src/main/java/com/pokemon/android/version/entity/pokemon/PokemonDataEntity.kt
@@ -4,6 +4,6 @@ data class PokemonDataEntity(
     val id: Int, val name: String, val type1: Int, val type2: Int,
     val possibleMoves: PossibleMovesEntity, val hp: Int, val attack: Int,
     val defense: Int, val spAtk: Int, val spDef: Int, val speed: Int,
-    val catchRate: Float, val evolutions : ArrayList<EvolutionEntity>,
+    val catchRate: Int, val evolutions : ArrayList<EvolutionEntity>,
     val expGaugeType : String) {
 }
